Update to Chromium 132.0.6834.83

+++ utils/clone.py
@@ -32,7 +32,6 @@ solutions = [
     "managed": False,
     "custom_deps": {
       "UC_OUT/third_party/angle/third_party/VK-GL-CTS/src": None,
-      "UC_OUT/third_party/instrumented_libs": None,
     },
     "custom_vars": {
       "checkout_configuration": "small",
@@ -237,22 +236,39 @@ def clone(args): # pylint: disable=too-many-branches, too-many-locals, too-many-
     # Match removals for the tarball:
     # https://source.chromium.org/chromium/chromium/tools/build/+/main:recipes/recipe_modules/chromium/resources/export_tarball.py
     remove_dirs = (
+        (args.output / 'base' / 'tracing' / 'test' / 'data'),
         (args.output / 'chrome' / 'test' / 'data'),
+        (args.output / 'components' / 'test' / 'data'),
         (args.output / 'content' / 'test' / 'data'),
         (args.output / 'courgette' / 'testdata'),
         (args.output / 'extensions' / 'test' / 'data'),
         (args.output / 'media' / 'test' / 'data'),
         (args.output / 'native_client' / 'src' / 'trusted' / 'service_runtime' / 'testdata'),
+        (args.output / 'testing' / 'libfuzzer' / 'fuzzers' / 'wasm_corpus'),
+        (args.output / 'third_party' / 'blink' / 'perf_tests'),
         (args.output / 'third_party' / 'blink' / 'tools'),
         (args.output / 'third_party' / 'blink' / 'web_tests'),
         (args.output / 'third_party' / 'breakpad' / 'breakpad' / 'src' / 'processor' / 'testdata'),
         (args.output / 'third_party' / 'catapult' / 'tracing' / 'test_data'),
+        (args.output / 'third_party' / 'dawn' / 'test'),
+        (args.output / 'third_party' / 'expat' / 'src' / 'testdata'),
+        (args.output / 'third_party' / 'harfbuzz-ng' / 'src' / 'test'),
         (args.output / 'third_party' / 'hunspell' / 'tests'),
         (args.output / 'third_party' / 'hunspell_dictionaries'),
         (args.output / 'third_party' / 'jdk' / 'current'),
         (args.output / 'third_party' / 'jdk' / 'extras'),
         (args.output / 'third_party' / 'liblouis' / 'src' / 'tests' / 'braille-specs'),
+        (args.output / 'third_party' / 'llvm' / 'llvm' / 'test'),
+        (args.output / 'third_party' / 'ots' / 'src' / 'tests' / 'fonts'),
+        (args.output / 'third_party' / 'rust-src' / 'src' / 'gcc' / 'gcc' / 'testsuite'),
+        (args.output / 'third_party' / 'rust-src' / 'src' / 'llvm-project' / 'clang' / 'test'),
+        (args.output / 'third_party' / 'rust-src' / 'src' / 'llvm-project' / 'llvm' / 'test'),
+        (args.output / 'third_party' / 'screen-ai' / 'linux' / 'resources'),
+        (args.output / 'third_party' / 'sqlite' / 'src' / 'test'),
+        (args.output / 'third_party' / 'swiftshader' / 'tests' / 'regres'),
+        (args.output / 'third_party' / 'test_fonts' / 'test_fonts'),
         (args.output / 'third_party' / 'xdg-utils' / 'tests'),
+        (args.output / 'tools' / 'perf' / 'testdata'),
         (args.output / 'v8' / 'test'),
     )
     keep_files = (
@@ -285,7 +301,8 @@ def clone(args): # pylint: disable=too-many-branches, too-many-locals, too-many-
                     path.rmdir()
     for path in sorted(args.output.rglob('*'), key=lambda l: len(str(l)), reverse=True):
         if not path.is_symlink() and '.git' not in path.parts:
-            if path.is_file() and ('out' in path.parts or path.name.startswith('ChangeLog')):
+            if path.is_file() and (('out' in path.parts and 'node_modules' not in path.parts)
+                                   or path.name.startswith('ChangeLog')):
                 try:
                     path.unlink()
                 except PermissionError:
